Refactor tile and tile node structures
- Removed the `Tile` struct from `src/models/tile/structs.rs`.
- Deleted the `tile_conversion.rs` file, which contained functions for converting grouped tile dimensions.
- Updated `TileDimensions` to change the type of `id` from `u8` to `i32` in `src/models/tile_dimensions/mod.rs`.
- Removed the `impls.rs` file for `TileNode` and integrated its content into `mod.rs`, enhancing the structure and organization.
- Deleted the `structs.rs` file for `TileNode`, consolidating the definition into `mod.rs`.
- Adjusted imports in `src/services/computation/mod.rs` to reflect the removal of the `performance_thresholds::structs` module.

+++ src/models/stock_solution/mod.rs
@@ -339,7 +339,7 @@ mod tests {
     use super::*;
     use crate::enums::orientation::Orientation;
 
-    fn create_test_tile(id: u8, width: u64, height: u64) -> TileDimensions {
+    fn create_test_tile(id: i32, width: u64, height: u64) -> TileDimensions {
         TileDimensions {
             id,
             width,
